requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Events==0.5
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opensearch-dsl==2.1.0
opensearch-py==2.8.0
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-multipart==0.0.19
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.44.0
